refactor: rename TX involved on enabling Prime on arbitrum
