--- multisig/simulations/arbitrumsepolia/vip-014/index.ts
@@ -1,109 +1,0 @@
-import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { expect } from "chai";
-import { Contract } from "ethers";
-import { parseUnits } from "ethers/lib/utils";
-import { ethers } from "hardhat";
-import { LzChainId } from "src/types";
-import { initMainnetUser } from "src/utils";
-import { forking, pretendExecutingVip } from "src/vip-framework";
-
-import vip014, {
-  MAX_DAILY_RECEIVE_LIMIT,
-  MAX_DAILY_SEND_LIMIT,
-  OP_SEPOLIA_TRUSTED_REMOTE,
-  SINGLE_RECEIVE_LIMIT,
-  SINGLE_SEND_LIMIT,
-} from "../../../proposals/arbitrumsepolia/vip-014";
-import XVS_ABI from "./abi/xvs.json";
-import XVS_BRIDGE_ABI from "./abi/xvsProxyOFTDest.json";
-
-const XVS = "0x877Dc896e7b13096D3827872e396927BbE704407";
-const XVS_BRIDGE = "0xFdC5cEC63FD167DA46cF006585b30D03B104eFD4";
-const XVS_HOLDER = "0xFd7dA20ea0bE63ACb0852f97E950376E7E4a817D";
-
-forking(79930177, async () => {
-  let xvs: Contract;
-  let xvsBridge: Contract;
-  let xvsHolderSigner: SignerWithAddress;
-  let receiver: SignerWithAddress;
-  let receiverAddressBytes32: string;
-  let defaultAdapterParams: string;
-
-  before(async () => {
-    xvs = await ethers.getContractAt(XVS_ABI, XVS);
-    xvsBridge = await ethers.getContractAt(XVS_BRIDGE_ABI, XVS_BRIDGE);
-    xvsHolderSigner = await initMainnetUser(XVS_HOLDER, ethers.utils.parseEther("5"));
-    [receiver] = await ethers.getSigners();
-    receiverAddressBytes32 = ethers.utils.defaultAbiCoder.encode(["address"], [receiver.address]);
-    defaultAdapterParams = ethers.utils.solidityPack(["uint16", "uint256"], [1, 300000]);
-  });
-
-  describe("Post-Execution state", () => {
-    before(async () => {
-      await pretendExecutingVip(await vip014());
-    });
-
-    it("Should match trusted remote address", async () => {
-      const trustedRemote = await xvsBridge.getTrustedRemoteAddress(LzChainId.opsepolia);
-      expect(trustedRemote).equals(OP_SEPOLIA_TRUSTED_REMOTE);
-    });
-
-    it("Should match single send transaction limit", async () => {
-      expect(await xvsBridge.chainIdToMaxSingleTransactionLimit(LzChainId.opsepolia)).to.equal(SINGLE_SEND_LIMIT);
-    });
-
-    it("Should match single receive transaction limit", async () => {
-      expect(await xvsBridge.chainIdToMaxSingleReceiveTransactionLimit(LzChainId.opsepolia)).to.equal(
-        SINGLE_RECEIVE_LIMIT,
-      );
-    });
-
-    it("Should match max daily send limit", async () => {
-      expect(await xvsBridge.chainIdToMaxDailyLimit(LzChainId.opsepolia)).to.equal(MAX_DAILY_SEND_LIMIT);
-    });
-
-    it("Should match max daily receive limit", async () => {
-      expect(await xvsBridge.chainIdToMaxDailyReceiveLimit(LzChainId.opsepolia)).to.equal(MAX_DAILY_RECEIVE_LIMIT);
-    });
-
-    it("Should emit an event on successful bridging of XVS (Arbitrum sepolia -> Op Sepolia)", async () => {
-      const amount = parseUnits("1", 18);
-      const nativeFee = (
-        await xvsBridge.estimateSendFee(
-          LzChainId.opsepolia,
-          receiverAddressBytes32,
-          amount,
-          false,
-          defaultAdapterParams,
-        )
-      ).nativeFee;
-
-      const circulatingSupplyBefore = await xvsBridge.circulatingSupply();
-      const totalSupplyBefore = await xvs.totalSupply();
-      const minterToMintedAmountBefore = await xvs.minterToMintedAmount(XVS_BRIDGE);
-
-      await expect(
-        xvsBridge
-          .connect(xvsHolderSigner)
-          .sendFrom(
-            xvsHolderSigner.address,
-            LzChainId.opsepolia,
-            receiverAddressBytes32,
-            amount,
-            [xvsHolderSigner.address, ethers.constants.AddressZero, defaultAdapterParams],
-            { value: nativeFee },
-          ),
-      )
-        .to.be.emit(xvsBridge, "SendToChain")
-        .withArgs(LzChainId.opsepolia, XVS_HOLDER, receiverAddressBytes32, amount);
-
-      const circulatingSupplyAfter = await xvsBridge.circulatingSupply();
-      const totalSupplyAfter = await xvs.totalSupply();
-      const minterToMintedAmountAfter = await xvs.minterToMintedAmount(XVS_BRIDGE);
-
-      expect(circulatingSupplyBefore).equals(circulatingSupplyAfter.add(amount));
-      expect(totalSupplyBefore).equals(totalSupplyAfter.add(amount));
-      expect(minterToMintedAmountBefore).equals(minterToMintedAmountAfter.add(amount));
-    });
-  });
-});
